add docs and rename file name
